nas-api/requirements.txt
flask==3.0.0
google-cloud-monitoring==2.21.0
